Refactoring

+++ LittleEngine/LittleEngine/Components/Mesh.h
@@ -22,7 +22,7 @@ struct Texture {
 	unsigned int id;
 	std::string type;
 	std::string path;
-	float textureSize;
+	unsigned int textureSize;
 };
 
 struct MeshInfo {
